bot/requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
distro==1.8.0
frozenlist==1.3.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
mypy==1.4.1
mypy-extensions==1.0.0
numpy==1.23.3
openai==1.3.2
pandas-stubs
peewee==3.16.0
Pillow==9.5.0
pydantic==2.5.1
pydantic_core==2.14.3
pytz==2023.3
requests==2.28.2
sniffio==1.3.0
tqdm==4.65.0
types-aiofiles==23.1.0.4
types-peewee==3.16.0.0
types-Pillow==9.5.0.5
types-pytz==2023.3.0.0
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.7.1
urllib3==1.26.15
yarl==1.8.2